+++ LangtonsAnt/logic.h
@@ -6,8 +6,7 @@
 #ifndef __LOGIC_H__
 #define __LOGIC_H__
 
-#include <limits.h>  // INT_MIN, INT_MAX
-#include <stddef.h>  // size_t
+#include <limits.h>
 
 
 /*--------------------- General purpose macros and types ---------------------*/
@@ -97,11 +96,11 @@ typedef signed char turn_t;
 
 /** Color rules container */
 typedef struct colors {
-	color_t  next[COLOR_COUNT];
-	turn_t   turn[COLOR_COUNT];
-	color_t  first, last;
-	color_t  def;
-	size_t   n;
+	color_t   next[COLOR_COUNT];
+	turn_t    turn[COLOR_COUNT];
+	color_t   first, last;
+	color_t   def;
+	unsigned  n;
 } Colors;  // TODO finish logic docs & add @see
 
 
@@ -137,7 +136,7 @@ typedef struct colors {
 
 /** Sparse matrix cell container */
 typedef struct cell {
-	size_t       packed;
+	unsigned     packed;
 	struct cell *next;
 } SparseCell;
 
@@ -146,8 +145,8 @@ typedef struct grid {
 	byte       **c, **tmp;
 	byte         def_color;
 	SparseCell **csr;
-	size_t       init_size, size, tmp_size;
-	size_t       colored;
+	unsigned     init_size, size, tmp_size;
+	unsigned     colored;
 	Vector2i     top_left, bottom_right;
 } Grid;
 
@@ -156,11 +155,11 @@ typedef struct grid {
 
 /** Simulation container */
 typedef struct simulation {
-	Colors *colors;
-	Grid   *grid;
-	Ant    *ant;
-	size_t  steps;
-	bool    is_running;
+	Colors   *colors;
+	Grid     *grid;
+	Ant      *ant;
+	unsigned  steps;
+	bool      is_running;
 } Simulation;
 
 
@@ -183,9 +182,9 @@ void colors_delete(Colors *colors);
 void colors_push(Colors *colors, color_t c, turn_t turn);
 void colors_pop(Colors *colors, color_t c);
 void colors_clear(Colors *colors);
-color_t colors_at(Colors *colors, size_t index);
-void colors_update(Colors *colors, size_t index, color_t c, turn_t turn);
-void colors_set_turn(Colors *colors, size_t index, turn_t turn);
+color_t colors_at(Colors *colors, unsigned index);
+void colors_update(Colors *colors, unsigned index, color_t c, turn_t turn);
+void colors_set_turn(Colors *colors, unsigned index, turn_t turn);
 bool color_exists(Colors *colors, color_t c);
 bool is_color_special(Colors *colors, color_t c);
 bool is_colors_empty(Colors *colors);
@@ -196,14 +195,14 @@ bool has_enough_colors(Colors *colors);
  *                                   grid.c                                   *
  *----------------------------------------------------------------------------*/
 
-Grid *grid_new(Colors *colors, size_t init_size);
+Grid *grid_new(Colors *colors, unsigned init_size);
 void grid_delete(Grid *grid);
 void grid_silent_expand(Grid *grid);
 void grid_expand(Grid *grid, Ant *ant);
 void grid_make_sparse(Grid *grid);
 bool is_grid_sparse(Grid *grid);
 bool is_grid_usage_low(Grid *grid);
-void sparse_new_cell(SparseCell **curr, size_t column, byte c);
+void sparse_new_cell(SparseCell **curr, unsigned column, byte c);
 byte sparse_color_at(Grid *grid, Vector2i p);
 
 
@@ -211,7 +210,7 @@ byte sparse_color_at(Grid *grid, Vector2i p);
  *                                simulation.c                                *
  *----------------------------------------------------------------------------*/
 
-Simulation *simulation_new(Colors *colors, size_t init_size);
+Simulation *simulation_new(Colors *colors, unsigned init_size);
 void simulation_delete(Simulation *sim);
 void simulation_run(Simulation *sim);
 void simulation_halt(Simulation *sim);
